Add package remove help and improve installation text
Currently the users do not know how to uninstall a package
once they installed it. Sadly, apturl does not provide an
option for this. And because I do not know what GUI the user
has for package management I simply use the command line.

Seperate the two installation options of the repository
so the users do not do them both.