Add uninstallation instructions
The installation description should also mention the
uninstallation in case users want to remove the repository
completely. This will not uninstall the packages but
there are instructions now on every app page.
Just add the uninstallation instruction for the package
version. Users who install it manually should also be
able to uninstall it manually.