Unduplicate the stats
There could be two packages which point to the same
application (by source and by package name).
e.g. avidemux2.6 and avidemux2.6-qt both point to "Avidemux 2.6".
As we need the package name only as a fallback when there is no
application name (which should be impossible anyway) do not
add it in the GROUP BY but just take the MAX() of it.
So only GROUP BY application name making it unique.